# Conflicts:
#	build.gradle
#	src/main/scala/beam/agentsim/agents/modalbehaviors/ChoosesMode.scala
#	src/main/scala/beam/agentsim/agents/modalbehaviors/DrivesVehicle.scala
#	src/main/scala/beam/agentsim/agents/ridehail/RideHailAgent.scala
#	src/main/scala/beam/agentsim/agents/ridehail/RideHailManager.scala
#	src/main/scala/beam/agentsim/events/handling/TravelTimeGoogleStatistic.scala
#	src/main/scala/beam/analysis/ModeChoiceAlternativesCollector.scala
#	src/main/scala/beam/router/BeamRouter.scala
#	src/main/scala/beam/router/TransitInitializer.scala
#	src/main/scala/beam/router/r5/R5Wrapper.scala
#	src/main/scala/beam/router/r5/RouteDumper.scala
#	src/main/scala/beam/router/skim/core/TransitCrowdingSkimmer.scala
#	src/main/scala/beam/router/skim/event/TransitCrowdingSkimmerEvent.scala
#	src/main/scala/beam/sim/BeamSim.scala
#	src/main/scala/beam/sim/BeamWarmStart.scala
#	src/main/scala/beam/sim/population/PopulationAdjustment.scala
#	src/main/scala/beam/sim/population/PopulationAttributes.scala
#	src/main/scala/beam/sim/population/PopulationScaling.scala
#	src/main/scala/beam/utils/MathUtils.scala
#	src/main/scala/beam/utils/data/ctpp/readers/BaseTableReader.scala
#	src/main/scala/beam/utils/data/synthpop/IndustryAssigner.scala
#	src/main/scala/beam/utils/data/synthpop/PopulationCorrection.scala
#	src/main/scala/beam/utils/data/synthpop/PopulationReader.scala
#	src/main/scala/beam/utils/data/synthpop/PumaLevelScenarioGenerator.scala
#	src/main/scala/beam/utils/data/synthpop/ScenarioGenerator.scala
#	src/main/scala/beam/utils/data/synthpop/generators/BenchmarkGenerator.scala
#	src/main/scala/beam/utils/map/R5MapStatsCalculator.scala
#	src/main/scala/beam/utils/scenario/generic/readers/PersonInfoReader.scala
#	src/main/scala/beam/utils/scenario/generic/readers/PlanElementReader.scala
#	src/main/scala/beam/utils/scenario/generic/writers/PersonInfoWriter.scala
#	src/main/scala/beam/utils/scenario/urbansim/CsvScenarioReader.scala
#	src/main/scala/beam/utils/scenario/urbansim/ParquetScenarioReader.scala
#	src/main/scala/beam/utils/scenario/urbansim/censusblock/merger/PersonMerger.scala
#	src/test/scala/beam/router/BeamRouterSpec.scala
#	src/test/scala/beam/router/skim/ODSkimmerTest.scala
#	src/test/scala/beam/router/skim/TransitCrowdingSkimsTest.scala
#	src/test/scala/beam/utils/scenario/urbansim/ParquetScenarioReaderTest.scala

+++ src/main/scala/beam/agentsim/agents/choice/mode/ModeChoiceLCCM.scala
@@ -214,7 +214,9 @@ class ModeChoiceLCCM(
       alternatives.zipWithIndex.map { altAndIdx =>
         val totalCost = altAndIdx._1.tripClassifier match {
           case TRANSIT | WALK_TRANSIT | DRIVE_TRANSIT | BIKE_TRANSIT =>
-            (altAndIdx._1.costEstimate + transitFareDefaults(altAndIdx._2)) * beamServices.beamConfig.beam.agentsim.tuning.transitPrice
+            (altAndIdx._1.costEstimate + transitFareDefaults(
+              altAndIdx._2
+            )) * beamServices.beamConfig.beam.agentsim.tuning.transitPrice
           case RIDE_HAIL =>
             altAndIdx._1.costEstimate * beamServices.beamConfig.beam.agentsim.tuning.rideHailPrice
           case _ =>
@@ -242,25 +244,24 @@ class ModeChoiceLCCM(
           walkTime,
           waitTime,
           bikeTime,
-          totalCost.toDouble,
+          totalCost,
           altAndIdx._2
         )
       }
 
     val groupedByMode: Map[BeamMode, Seq[ModeChoiceData]] =
       modeChoiceAlternatives.groupBy(_.mode)
-    val bestInGroup = groupedByMode.map {
-      case (_, alts) =>
-        // Which dominates at $18/hr for total time
-        alts
-          .map { alt =>
-            (
-              (alt.vehicleTime + alt.walkTime + alt.waitTime + alt.bikeTime) / 3600 * 18 + alt.cost,
-              alt
-            )
-          }
-          .minBy(_._1)
-          ._2
+    val bestInGroup = groupedByMode.map { case (_, alts) =>
+      // Which dominates at $18/hr for total time
+      alts
+        .map { alt =>
+          (
+            (alt.vehicleTime + alt.walkTime + alt.waitTime + alt.bikeTime) / 3600 * 18 + alt.cost,
+            alt
+          )
+        }
+        .minBy(_._1)
+        ._2
     }
     bestInGroup.toVector
   }
@@ -301,9 +302,8 @@ class ModeChoiceLCCM(
         )
       }
       .toMap
-      .mapValues(
-        modeChoiceCalculatorForStyle =>
-          trips.map(trip => modeChoiceCalculatorForStyle.utilityOf(trip, attributesOfIndividual, None)).sum
+      .mapValues(modeChoiceCalculatorForStyle =>
+        trips.map(trip => modeChoiceCalculatorForStyle.utilityOf(trip, attributesOfIndividual, None)).sum
       )
       .toArray
       .toMap // to force computation DO NOT TOUCH IT, because here is call-by-name and it's lazy which will hold a lot of memory !!! :)
@@ -314,17 +314,16 @@ class ModeChoiceLCCM(
     val logsum = Option(
       math.log(
         person.getPlans.asScala.view
-          .map(
-            plan =>
-              plan.getAttributes
-                .getAttribute("scores")
-                .asInstanceOf[MapStringDouble]
-                .data(attributesOfIndividual.modalityStyle.get)
+          .map(plan =>
+            plan.getAttributes
+              .getAttribute("scores")
+              .asInstanceOf[MapStringDouble]
+              .data(attributesOfIndividual.modalityStyle.get)
           )
           .map(score => math.exp(score))
           .sum
       )
-    ).filterNot(x => x < -100D).getOrElse(-100D)
+    ).filterNot(x => x < -100d).getOrElse(-100d)
 
     // Score of being in class given this outcome
     lccm
